web/DJANGO_K8S/wsgi.py
"""
WSGI config for DJANGO_K8S project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/4.2/howto/deployment/wsgi/
"""

import os
import pathlib
import dotenv

from django.core.wsgi import get_wsgi_application

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'DJANGO_K8S.settings')

application = get_wsgi_application()
